Rename test app folder
