+++ e2e/browser/testApp/next-env.d.ts
