+++ e2e/browser/testApp/next.config.js
