+++ e2e/browser/testApp/package.json
