+++ e2e/browser/testApp/tsconfig.json
